# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SailPoint RPA Studio",
    "description": "Phase 1 — web RPA orchestration. SailPoint adapter is planned for phase 3.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/applications": {
      "get": {
        "tags": [
          "Applications"
        ],
        "summary": "List Applications",
        "operationId": "list_applications_api_v1_applications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Applications"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_applications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/workflows": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "List Workflows",
        "operationId": "list_workflows_api_v1_workflows_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Workflows"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_workflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/workflows/{workflow_id}": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "Get Workflow",
        "operationId": "get_workflow_api_v1_workflows__workflow_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Workflows"
        ],
        "summary": "Save",
        "operationId": "save_api_v1_workflows__workflow_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplicationIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "url": {
            "type": "string",
            "maxLength": 2000,
            "title": "Url"
          },
          "environment": {
            "type": "string",
            "enum": [
              "DEV",
              "HML",
              "PRD"
            ],
            "title": "Environment",
            "default": "DEV"
          },
          "browser": {
            "type": "string",
            "enum": [
              "chromium",
              "firefox",
              "chrome",
              "msedge"
            ],
            "title": "Browser",
            "default": "chromium"
          },
          "headless": {
            "type": "boolean",
            "title": "Headless",
            "default": true
          },
          "timeout_ms": {
            "type": "integer",
            "maximum": 60000.0,
            "minimum": 500.0,
            "title": "Timeout Ms",
            "default": 30000
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Tags"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "url"
        ],
        "title": "ApplicationIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginIn": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginIn"
      },
      "Selector": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "css",
              "testid",
              "role",
              "label",
              "placeholder",
              "text",
              "xpath"
            ],
            "title": "Kind",
            "default": "css"
          },
          "value": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Value"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "value"
        ],
        "title": "Selector"
      },
      "Step": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,60}$",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "navigate",
              "click",
              "fill",
              "wait",
              "select",
              "screenshot",
              "extract_text",
              "assert_text",
              "check",
              "press",
              "hover"
            ],
            "title": "Type"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "title": "Name",
            "default": ""
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "selectors": {
            "items": {
              "$ref": "#/components/schemas/Selector"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Selectors"
          },
          "value": {
            "type": "string",
            "maxLength": 10000,
            "title": "Value",
            "default": ""
          },
          "url": {
            "type": "string",
            "maxLength": 2000,
            "title": "Url",
            "default": ""
          },
          "output": {
            "type": "string",
            "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,59}$",
            "title": "Output",
            "default": "result"
          },
          "timeout_ms": {
            "type": "integer",
            "maximum": 60000.0,
            "minimum": 200.0,
            "title": "Timeout Ms",
            "default": 10000
          },
          "wait_ms": {
            "type": "integer",
            "maximum": 30000.0,
            "minimum": 0.0,
            "title": "Wait Ms",
            "default": 1000
          },
          "secret": {
            "type": "boolean",
            "title": "Secret",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type"
        ],
        "title": "Step"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkflowIn": {
        "properties": {
          "application_id": {
            "type": "string",
            "title": "Application Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "operation": {
            "type": "string",
            "pattern": "^[A-Z_]{1,40}$",
            "title": "Operation",
            "default": "CUSTOM"
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 5.0,
            "title": "Timeout Seconds",
            "default": 600
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/Step"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Steps"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "application_id",
          "name"
        ],
        "title": "WorkflowIn"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}